requirements.txt
# this file lists all the dependencies needed for running heat

d2to1>=0.2.10,<0.3
pbr>=0.5.10,<0.6

# The greenlet package must be compiled with gcc and needs
# the Python.h headers. Make sure you install the python-dev
# package to get the right headers...

# lxml note: You will need gcc buildtools installed and must
# have installed libxml headers for lxml to be successfully
# installed using pip, therefore you will need to install the
# libxml2-dev(el) and libxslt-dev(el) packages.

PyCrypto>=2.6
boto>=2.4.0
eventlet>=0.12.0
greenlet>=0.3.2
httplib2
iso8601>=0.1.4
kombu>=1.0.4
argparse
lxml>=2.3
six
sqlalchemy-migrate>=0.7.2
python-novaclient>=2.11.0,<3
PasteDeploy==1.5.0
routes==1.12.3
SQLAlchemy>=0.7.8,<0.7.99
WebOb==1.2.3
python-keystoneclient>=0.2.3
python-memcached
python-swiftclient
python-quantumclient>=2.2.0,<3.0.0
python-cinderclient
PyYAML>=3.1.0
oslo.config>=1.1.0
paramiko
pyrax
